# server/fleetv2_http_api/models/message.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from fleetv2_http_api.models.base_model import Model
from fleetv2_http_api.models.device_id import DeviceId
from fleetv2_http_api.models.payload import Payload
from fleetv2_http_api import util

from fleetv2_http_api.models.device_id import DeviceId  # noqa: E501
from fleetv2_http_api.models.payload import Payload  # noqa: E501

class Message(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, timestamp=None, id=None, payload=None):  # noqa: E501
        """Message - a model defined in OpenAPI

        :param timestamp: The timestamp of this Message.  # noqa: E501
        :type timestamp: int
        :param id: The id of this Message.  # noqa: E501
        :type id: DeviceId
        :param payload: The payload of this Message.  # noqa: E501
        :type payload: Payload
        """
        self.openapi_types = {
            'timestamp': int,
            'id': DeviceId,
            'payload': Payload
        }

        self.attribute_map = {
            'timestamp': 'timestamp',
            'id': 'id',
            'payload': 'payload'
        }

        self._timestamp = timestamp
        self._id = id
        self._payload = payload

    @classmethod
    def from_dict(cls, dikt) -> 'Message':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Message of this Message.  # noqa: E501
        :rtype: Message
        """
        return util.deserialize_model(dikt, cls)

    @property
    def timestamp(self) -> int:
        """Gets the timestamp of this Message.


        :return: The timestamp of this Message.
        :rtype: int
        """
        return self._timestamp

    @timestamp.setter
    def timestamp(self, timestamp: int):
        """Sets the timestamp of this Message.


        :param timestamp: The timestamp of this Message.
        :type timestamp: int
        """

        self._timestamp = timestamp

    @property
    def id(self) -> DeviceId:
        """Gets the id of this Message.


        :return: The id of this Message.
        :rtype: DeviceId
        """
        return self._id

    @id.setter
    def id(self, id: DeviceId):
        """Sets the id of this Message.


        :param id: The id of this Message.
        :type id: DeviceId
        """

        self._id = id

    @property
    def payload(self) -> Payload:
        """Gets the payload of this Message.


        :return: The payload of this Message.
        :rtype: Payload
        """
        return self._payload

    @payload.setter
    def payload(self, payload: Payload):
        """Sets the payload of this Message.


        :param payload: The payload of this Message.
        :type payload: Payload
        """

        self._payload = payload